+++ sdk/src/main/java/com/mapbox/maps/viewannotation/ViewAnnotationManager.kt
@@ -144,4 +144,25 @@ interface ViewAnnotationManager {
    * Remove an instance of [OnViewAnnotationUpdatedListener].
    */
   fun removeOnViewAnnotationUpdatedListener(listener: OnViewAnnotationUpdatedListener)
+
+  /**
+   * Set view annotation update mode defined in [ViewAnnotationUpdateMode].
+   */
+  fun setViewAnnotationUpdateMode(mode: ViewAnnotationUpdateMode)
+
+  /**
+   * Get view annotation update mode.
+   */
+  fun getViewAnnotationUpdateMode(): ViewAnnotationUpdateMode
+
+  /**
+   * Static methods and variables.
+   */
+  companion object {
+    /**
+     * Default view annotation update mode.
+     */
+    @JvmField
+    val DEFAULT_UPDATE_MODE = ViewAnnotationUpdateMode.MAP_SYNCHRONIZED
+  }
 }